iaasApi permission factory

+++ src/main/java/io/mixeway/rest/project/service/IaasApiService.java
@@ -3,6 +3,7 @@ package io.mixeway.rest.project.service;
 import io.mixeway.db.entity.IaasApiType;
 import io.mixeway.db.entity.Project;
 import io.mixeway.db.repository.IaasApiTypeRepisotory;
+import io.mixeway.pojo.PermissionFactory;
 import io.mixeway.rest.project.model.IaasApiPutModel;
 import io.mixeway.rest.project.model.IaasModel;
 import org.slf4j.Logger;
@@ -16,6 +17,7 @@ import io.mixeway.db.repository.ProjectRepository;
 import io.mixeway.pojo.Status;
 
 import javax.transaction.Transactional;
+import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
 
@@ -25,20 +27,25 @@ public class IaasApiService {
     private final ProjectRepository projectRepository;
     private final IaasApiRepository iaasApiRepository;
     private final IaasApiTypeRepisotory iaasApiTypeRepisotory;
+    private final PermissionFactory permissionFactory;
     private final io.mixeway.integrations.servicediscovery.service.IaasService iaasApiService;
 
 
     IaasApiService(ProjectRepository projectRepository, IaasApiTypeRepisotory iaasApiTypeRepisotory,
-                   IaasApiRepository iaasApiRepository, io.mixeway.integrations.servicediscovery.service.IaasService iaasApiService){
+                   IaasApiRepository iaasApiRepository, io.mixeway.integrations.servicediscovery.service.IaasService iaasApiService,
+                   PermissionFactory permissionFactory){
         this.projectRepository = projectRepository;
         this.iaasApiRepository = iaasApiRepository;
         this.iaasApiTypeRepisotory = iaasApiTypeRepisotory;
         this.iaasApiService = iaasApiService;
+        this.permissionFactory = permissionFactory;
     }
 
-    public ResponseEntity<IaasModel> showIaasApi(Long id) {
+    public ResponseEntity<IaasModel> showIaasApi(Long id, Principal principal) {
         Optional<Project> project = projectRepository.findById(id);
-        if ( project.isPresent() && project.get().getIaasApis().size() == 1){
+        if ( project.isPresent() &&
+                permissionFactory.canUserAccessProject(principal, project.get()) &&
+                project.get().getIaasApis().size() == 1){
             Optional<IaasApi> iaasApi = project.get().getIaasApis().stream().findFirst();
             IaasModel iaasModel = new IaasModel();
             if (iaasApi.isPresent()) {
@@ -58,14 +65,14 @@ public class IaasApiService {
         }
     }
 
-    public ResponseEntity<Status> saveIaasApi(Long id, IaasApiPutModel iaasApiPutModel, String username) {
+    public ResponseEntity<Status> saveIaasApi(Long id, IaasApiPutModel iaasApiPutModel, Principal principal) {
         Optional<Project> project = projectRepository.findById(id);
-        if (project.isPresent()){
+        if (project.isPresent() && permissionFactory.canUserAccessProject(principal, project.get())){
             if (project.get().getIaasApis().size() >0){
                 return new ResponseEntity<>(null,HttpStatus.EXPECTATION_FAILED);
             } else {
                iaasApiService.saveApi(iaasApiPutModel,project.get());
-                log.info("{} - Saved new IaasApi for project {}", username, project.get().getName());
+                log.info("{} - Saved new IaasApi for project {}", principal.getName(), project.get().getName());
                 return new ResponseEntity<>(new Status("ok"), HttpStatus.CREATED);
             }
         } else {
@@ -73,10 +80,10 @@ public class IaasApiService {
         }
     }
 
-    public ResponseEntity<Status> testIaasApi(Long id) {
+    public ResponseEntity<Status> testIaasApi(Long id, Principal principal) {
 
         Optional<Project> project = projectRepository.findById(id);
-        if (project.isPresent()){
+        if (project.isPresent() && permissionFactory.canUserAccessProject(principal, project.get())){
             Optional<IaasApi> api = project.get().getIaasApis().stream().findFirst();
             if (api.isPresent() && api.get().getProject().getId().equals(project.get().getId())) {
                 try {
@@ -93,14 +100,14 @@ public class IaasApiService {
         return new ResponseEntity<>(null,HttpStatus.EXPECTATION_FAILED);
     }
 
-    public ResponseEntity<Status> iaasApiEnableSynchro(Long id, String username) {
+    public ResponseEntity<Status> iaasApiEnableSynchro(Long id, Principal principal) {
         Optional<Project> project = projectRepository.findById(id);
-        if (project.isPresent()){
+        if (project.isPresent() && permissionFactory.canUserAccessProject(principal, project.get())){
             Optional<IaasApi> api = project.get().getIaasApis().stream().findFirst();
             if (api.isPresent() && api.get().getProject().getId().equals(project.get().getId()) && api.get().getStatus()) {
                 api.get().setEnabled(true);
                 iaasApiRepository.save(api.get());
-                log.info("{} - Enabled auto synchro of IaasApi for project {}", username, project.get().getName());
+                log.info("{} - Enabled auto synchro of IaasApi for project {}", principal.getName(), project.get().getName());
                 return new ResponseEntity<>(null,HttpStatus.OK);
             }
         } else {
@@ -109,14 +116,14 @@ public class IaasApiService {
         return new ResponseEntity<>(null,HttpStatus.EXPECTATION_FAILED);
     }
 
-    public ResponseEntity<Status> iaasApiDisableSynchro(Long id, String username) {
+    public ResponseEntity<Status> iaasApiDisableSynchro(Long id, Principal principal) {
         Optional<Project> project = projectRepository.findById(id);
-        if (project.isPresent()){
+        if (project.isPresent() && permissionFactory.canUserAccessProject(principal, project.get())){
             Optional<IaasApi> api = project.get().getIaasApis().stream().findFirst();
             if (api.isPresent() && api.get().getProject().getId().equals(project.get().getId())) {
                 api.get().setEnabled(false);
                 iaasApiRepository.save(api.get());
-                log.info("{} - Disabled auto synchro of IaasApi for project {}", username, project.get().getName());
+                log.info("{} - Disabled auto synchro of IaasApi for project {}", principal.getName(), project.get().getName());
                 return new ResponseEntity<>(null,HttpStatus.OK);
             }
         } else {
@@ -125,13 +132,13 @@ public class IaasApiService {
         return new ResponseEntity<>(null,HttpStatus.EXPECTATION_FAILED);
     }
     @Transactional
-    public ResponseEntity<Status> iaasApiDelete(Long id, String username) {
+    public ResponseEntity<Status> iaasApiDelete(Long id, Principal principal) {
         Optional<Project> project = projectRepository.findById(id);
-        if (project.isPresent()){
+        if (project.isPresent() && permissionFactory.canUserAccessProject(principal, project.get())){
             Optional<IaasApi> api = project.get().getIaasApis().stream().findFirst();
             if (api.isPresent() && api.get().getProject() == project.get()) {
                 iaasApiRepository.delete(api.get());
-                log.info("{} - deleted IaasApi for project {}", username, project.get().getName());
+                log.info("{} - deleted IaasApi for project {}", principal.getName(), project.get().getName());
                 return new ResponseEntity<>(null,HttpStatus.OK);
             }
         } else {
@@ -140,7 +147,7 @@ public class IaasApiService {
         return new ResponseEntity<>(null,HttpStatus.EXPECTATION_FAILED);
     }
 
-    public ResponseEntity<List<IaasApiType>> getIaasApiTypes(String name) {
+    public ResponseEntity<List<IaasApiType>> getIaasApiTypes(Principal principal) {
         return new ResponseEntity<>(iaasApiTypeRepisotory.findAll(),HttpStatus.OK);
     }
 }
